Progress
The web app now pulls data from github and Fullstory using their
REST APIs, in order to list all the open issues for the calculator web app,
and list the username, link to the github issue, and link to Fullstory session
for each user who submitted a github issue.

Problem: I still need to pull the email of the github user, to tie to Fullstory
sessions. Using a hard-coded email right now.